Maximum and minimum facets can't handle > 32-bit numbers
Fixes #280

+++ raml-parser-2/src/main/java/org/raml/v2/api/model/v08/parameters/StringTypeDeclaration.java
@@ -42,6 +42,6 @@ public interface StringTypeDeclaration extends Parameter
     /**
      * (Optional, applicable only for parameters of type string) The maxLength attribute specifies the parameter value's maximum number of characters.
      **/
-    Integer maxLength();
+    Long maxLength();
 
 }
