+++ raml-parser-2/src/main/java/org/raml/v2/api/model/v10/datamodel/ArrayTypeDeclaration.java
@@ -34,12 +34,12 @@ public interface ArrayTypeDeclaration extends TypeDeclaration
     /**
      * Minimum amount of items in array
      **/
-    Integer minItems();
+    Long minItems();
 
 
     /**
      * Maximum amount of items in array
      **/
-    Integer maxItems();
+    Long maxItems();
 
 }
