+++ raml-parser-2/src/main/java/org/raml/v2/api/model/v10/datamodel/StringTypeDeclaration.java
@@ -30,13 +30,13 @@ public interface StringTypeDeclaration extends TypeDeclaration
     /**
      * Minimum length of the string
      **/
-    Integer minLength();
+    Long minLength();
 
 
     /**
      * Maximum length of the string
      **/
-    Integer maxLength();
+    Long maxLength();
 
 
     /**
