+++ yagi/src/main/java/org/raml/yagi/framework/grammar/rule/MaxPropertiesRule.java
@@ -25,9 +25,9 @@ import java.util.List;
 
 public class MaxPropertiesRule extends Rule
 {
-    private int maxProperties;
+    private long maxProperties;
 
-    public MaxPropertiesRule(int maxProperties)
+    public MaxPropertiesRule(long maxProperties)
     {
         this.maxProperties = maxProperties;
     }
