+++ yagi/src/main/java/org/raml/yagi/framework/model/SimpleValueTransformer.java
@@ -90,7 +90,7 @@ public enum SimpleValueTransformer
         }
 
         @Override
-        public Integer adaptTo(Node valueNode, Class<?> returnType)
+        public Long adaptTo(Node valueNode, Class<?> returnType)
         {
             if (NodeUtils.isNull(valueNode))
             {
@@ -102,7 +102,7 @@ public enum SimpleValueTransformer
             }
             if (valueNode instanceof SimpleTypeNode)
             {
-                return Integer.parseInt(((SimpleTypeNode) valueNode).getLiteralValue());
+                return Long.parseLong(((SimpleTypeNode) valueNode).getLiteralValue());
             }
             else
             {
